fixed failing tests
